refactor: cria diretório para os tipos do AS
