+++ types/plunder/attack.ts
@@ -1,0 +1,8 @@
+export type PlunderAttackDetails = {
+    wood: number;
+    stone: number;
+    iron: number;
+    total: number;
+    attackAmount: number;
+    destroyedWalls: number;
+};
