+++ types/plunder/history.ts
@@ -1,0 +1,6 @@
+import type { PlunderAttackDetails } from '$types/plunder/attack';
+
+export type PlunderHistoryType = {
+    last: PlunderAttackDetails;
+    total: PlunderAttackDetails;
+};
