+++ types/plunder/index.ts
@@ -1,0 +1,7 @@
+export type * from '$types/plunder/attack';
+export type * from '$types/plunder/config';
+export type * from '$types/plunder/group';
+export type * from '$types/plunder/history';
+export type * from '$types/plunder/info';
+export type * from '$types/plunder/page';
+export type * from '$types/plunder/other';
